Generate Filestore in Terraform in Magic Modules.
Two significant magic modules changes here:

* This change adds the new parameter setting 'url_param_only',
which suppresses the value in the body of HTTP requests.  This is
necessary in new-style products, where /location/ is used instead
of zone or region.

* This change fixes an issue with 'ignore_read', which previously
only worked on top-level fields.  With this change, nested fields
also are not set on read if marked 'ignore_read'.

Both of these are implemented in terraform only.