Do not send empty values for nested properties
Top level terraform properties do are removed from the api payload if they are
not set in the terraform config. However this behavior is different for any
properties nested past the top level. This results in empty objects getting
sent in payloads where every sub property has been defaulted to its empty state.
This causes problems for more complicated apis that have deeper nesting such as
the stack driver apis.

This change will remove any property from the payload if it is not defined in
the terraform config.