Http check auth work arounds
http check supports basic auth by passing in a username and password. However
the api will return "*****" for password if the password has been set. This
makes it difficult to determine if the password has been changed since the
response is always the same string. As a work around I've set a different
field with the actual content of the password on create and udpate so that it
can be checked against the value in the plan to determine if there are changes
that need to be sent.