Temporarily disable SQL sweepers in Terraform.
Terraform's test sweepers run before every CI run, and fail the CI run
if there's an error deleting things. Right now, we have an SQL instance
that stubbornly refuses to be deleted, and it's breaking CI. As we work
with support to resolve that situation, let's disable the sweeper so we
can run the CI again.