Add support for Audit Configs.
This adds the ability to manage audit log configs with the
google_iam_policy data source and the google_project_iam_policy
resource. A few notes about this implementation:

1. Users that have manually turned on audit configs and use the
google_project_iam_policy resource will have their audit configs
deleted. This should have been happening anyways, so I don't see that as
a big deal.

2. To manage audit configs with Terraform after this implementation, you
_must_ use the google_project_iam_policy resource, which is less than
ideal. A next step (which hopefully I'll achieve this week) is to add
audit configs to our IAM framework, so all of the resources we support
IAM policies, bindings, and members for will now also support audit
configs. This is a more reasonable user experience, allowing users to
add audit configs with more fine-grained control. The only thing to
investigate is whether it makes sense as part of the framework, or as a
one-off for projects only, which largely depends on whether audit
configs can only be used on project IAM policies or on IAM policies for
individual resources. I'll investigate that more.