Ordering changes in create-prs.
Ordering is extremely important to the magician.  To save time, we don't
close the repo 'magic-modules' back in - but we need it later on to
run jobs out of.  This crashes the magician.  Also, the 'with-comment'
repo does not exist at the time that we are trying to use it to run
jobs.