Improve to_json output
For deeply nested resources the existence of the `@properties` object was
making it hard to read and doubling the nesting. This will remove that
special property and just surface the properties as JSON properties on the
object.

Variables for a given field will be JSON properties up with the @var name
while child properties will not have the @ prefix.