BigQuery: table filtering/clustering/partitioning support.
Intent here is to add support for specifying the column for
partitioned tables, allowing specification of secondary data
clustering, and enabling the ability to designate the ability
to disallow unbounded full table scans on large partitioned
tables.

That said, while I'm familiar with the BQ API I'm very unfamiliar
with this tooling, so this may be a glorious example of holding the
tools wrong.