Catch 404s when deleting default networks.
As seen in terraform-providers/terraform-provider-google#3582, it is now
possible to set an organization policy that removes the default network
from a project when it's created. This means it's now possible that
Terraform's attempt to delete that default network will encounter an
error saying the network is not found. Because what Terraform wanted was
achieved, even if not by Terraform, we shouldn't raise that error, we
should ignore it.