Unbreak IAM.
I tried to reintroduce case-insensitivity for members in IAM in #2168,
but apparently only _part_ of the members value is case-insensitive.

Members values are in the format `<type>:<value>` where `<type>` is
something like `user` and `<value>` is something like
`paddy@hashicorp.com`. The _value_ is case-insensitive, but the _type_
is very much case-sensitive, and broke a bunch of our tests, namely for
the `serviceAccount` type.

This fixes it by only changing the case of the value, not the whole
string.