Fixes discovered in testing GCB/GH differ.
The GCB / GH integration doesn't satisfy our use case perfectly.
It doesn't check out the merge commit, and it doesn't check out the repo
itself - it only gives us the actual *code*, not the repo.  So we need
to handle that access ourselves in the beginning.  We also need a
beefier machine to make this work in a reasonable amount of time.